meta-agl-bsp: remove linux-raspberrypi_5.4.bbappend
Remove the Xen-specific linux-raspberry kernel bbappend and associated
patch as they are no longer applicable to the default 5.10 kernel.
Based on Xen project announcements, it is believed that Xen should
work out of the box with 5.10 on Raspberry Pi 4, but that has not been
tested as part of this upgrade.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I879f9ddb63f1ad83051df56f167d63c901a7b5e8