pipewire: remove old meson compatibility patches
Remove accumulated patches to PipeWire that allowed building it with
the older meson in dunfell, as they break building it newer meson
in poky master/kirkstone.  A fix contained in one of those patches
for installing the systemd system unit has been reworked and added
as a new separate patch.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I949875339b500c9ccd8590a635f7734158617579